tools/document-converter/requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0
